# Implemented ActorRef caching in Akka.Remote