--- a/src/core/Akka.Remote.Tests/Akka.Remote.Tests.csproj
+++ b/src/core/Akka.Remote.Tests/Akka.Remote.Tests.csproj
@@ -94,6 +94,7 @@
   <ItemGroup>
     <Compile Include="AccrualFailureDetectorSpec.cs" />
     <Compile Include="AckedDeliverySpec.cs" />
+    <Compile Include="ActorRefCacheSpec.cs" />
     <Compile Include="DeadlineFailureDetectorSpec.cs" />
     <Compile Include="EndpointRegistrySpec.cs" />
     <Compile Include="FailureDetectorRegistrySpec.cs" />