--- a/src/core/Akka.Remote/Configuration/Remote.conf
+++ b/src/core/Akka.Remote/Configuration/Remote.conf
@@ -90,6 +90,12 @@ akka {
     # every module will respect this setting.
     use-dispatcher = "akka.remote.default-remote-dispatcher"
 
+    actorref-cache {
+        enabled = true
+        maximum-remote-cache-size = 4096
+        remote-cache-expiration = 30s
+    }
+
     ### Security settings
 
     # Enable untrusted mode for full security of server managed actors, prevents