This PR mitigates the current bottlenecks in Akka.Remote:
- ActorPath parsing
- ActorRef resolving

The cache is backed by a ConcurrentDictionary which maps the string path straight to the ActorRef. An actor is responsible for the handling of cache maximum size and expiration with an LRU list.

There is two modes of operation:
#####  Local ActorRefWithCell caching
This mode handles the caching of local actor refs backed with an ActorCell. It is bounded by the number of live actors in the system. The cache entries are cleared with DeathWatch notifications.

##### Remote ActorRef caching
This mode caches all the actor refs not backed with an ActorCell (e.g. RemoteActorRef). It is bounded by the `maximum-remote-cache-size` setting, the least recently used actor refs are removed first when the maximum capacity is reached.

The cache entries in this mode are also cleared by an expiration specified with the `remote-cache-expiration` setting.

##### Note
Actor refs under the /temp path are not cached since they will most likely not be reused. This means that Asks will not benefit from this optimization. A workaround would be to set the `Sender` of an Ask to a system actor that would forward the response to the FutureActorRef.