@JeffCyr 2016-05-24
On my laptop, I get theses results with the RemoteBenchmark:

##### Without caching:
Optimized Echo without sender: 70K msg/sec
Echo with sender: 50K msg/sec

##### With caching
Optimized Echo without sender: 118K msg/sec
Echo with sender: 100K msg/sec

======
#### NBench remote throughput increased from 15,686.02 msg/sec to 27,006.78 msg/sec.

# Without caching:
# Akka.Remote.Tests.Performance.ForkJoinDispatcherRemoteMessagingThroughputSpec+TwoWay
__Measures the throughput of Akka.Remote over a particular transport using two-way messaging__
_5/24/2016 9:21:37 AM_
### System Info
```ini
NBench=NBench, Version=0.2.1.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           57.00 |           57.00 |           57.00 |            0.00 |
TotalCollections [Gen1] |     collections |            5.00 |            3.46 |            2.00 |            0.78 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |       10,000.00 |       10,000.00 |       10,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           96.51 |           89.41 |           80.64 |            5.52 |
TotalCollections [Gen1] |     collections |            7.82 |            5.41 |            3.09 |            1.17 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |       16,930.80 |       15,686.02 |       14,147.66 |          968.12 |

========

# With caching:
# Akka.Remote.Tests.Performance.ForkJoinDispatcherRemoteMessagingThroughputSpec+TwoWay
__Measures the throughput of Akka.Remote over a particular transport using two-way messaging__
_5/24/2016 10:54:47 AM_
### System Info
```ini
NBench=NBench, Version=0.2.1.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           37.00 |           37.00 |           37.00 |            0.00 |
TotalCollections [Gen1] |     collections |            2.00 |            1.38 |            1.00 |            0.51 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |       10,000.00 |       10,000.00 |       10,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          107.62 |           99.93 |           95.25 |            3.88 |
TotalCollections [Gen1] |     collections |            5.47 |            3.72 |            2.57 |            1.32 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |       29,085.74 |       27,006.78 |       25,743.38 |        1,047.47 |