@rogeralsing 2016-05-24
Looking good to me, can't see anything to comment on code-wise.